feat: Initial SDL3 and WGPU integration (hardcoded triangle). (#43)
* Initial SDL3 and WGPU integration (hardcoded triangle).

* Enabled NativeAOT compilation.

* Added configuration support for Mac and Browser (both untested).

* Implemented indexed quad rendering.

* Hot reload and multiple render stage support.

* Updated solution (still WIP).

* Reverted enabling of WGPU

* Fixed scene example project.

+++ Ion/Ion.Extensions.Audio/Assets/SoundEffect.cs
@@ -6,7 +6,7 @@ namespace Ion.Extensions.Audio
 {
 	public class SoundEffect : ISoundEffect
 	{
-		public int Id { get; }
+		public nint Id { get; }
 
 		public string Name { get; }
